requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
fastapi==0.116.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
pydantic[email]==2.11.9
pydantic_core==2.33.2
passlib[bcrypt]==1.7.4
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-xprocess==1.0.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47
twilio==9.8.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
razorpay==1.4.2
